fix: extract `InputCheckboxGroupItem` to be also used in `InputResourceGroup`
