+++ packages/app-elements/src/ui/forms/InputCheckboxGroup/index.tsx
@@ -1,0 +1,4 @@
+export {
+  InputCheckboxGroup,
+  type InputCheckboxGroupProps
+} from './InputCheckboxGroup'
