+++ packages/app-elements/src/ui/forms/InputResourceGroup/utils.ts
@@ -1,3 +1,5 @@
+import { type CommerceLayerClient } from '@commercelayer/sdk'
+import { type ListableResourceType } from '@commercelayer/sdk/lib/cjs/api'
 import { useState } from 'react'
 
 export function useToggleCheckboxValues(defaultValues: string[]): {
@@ -48,4 +50,40 @@ export function computeLabelWithSelected({
   const counter =
     selectedCount > 0 ? `${selectedCount} of ${totalCount}` : totalCount
   return `${label} · ${counter}`
+}
+
+type Resource = Awaited<
+  ReturnType<CommerceLayerClient[ListableResourceType]['list']>
+>[number]
+
+/**
+ * Helper to prepare the options for the checkbox group or mocked info to be used for skeleton.
+ */
+export function prepareCheckboxItemOrMock({
+  resource,
+  isLoading,
+  fieldForLabel,
+  fieldForValue
+}: {
+  resource?: Resource
+  isLoading?: boolean
+  fieldForValue: string
+  fieldForLabel: string
+}): { value: string; label: string } {
+  const getAttribute = (attribute: string): string =>
+    resource == null
+      ? ''
+      : ((resource[attribute as keyof Resource] ?? '') as string)
+
+  return isLoading === true || resource == null
+    ? // mock
+      {
+        value: '',
+        label: 'Commerce Layer'
+      }
+    : // real
+      {
+        value: getAttribute(fieldForValue),
+        label: getAttribute(fieldForLabel)
+      }
 }
