+++ packages/docs/src/stories/forms/ui/InputCheckboxGroup.stories.tsx
@@ -1,6 +1,5 @@
 import { Avatar } from '#ui/atoms/Avatar'
 import { Text } from '#ui/atoms/Text'
-import { ListItem } from '#ui/composite/ListItem'
 import { InputCheckboxGroup } from '#ui/forms/InputCheckboxGroup'
 import { type Meta, type StoryFn } from '@storybook/react'
 import { useState } from 'react'
@@ -35,92 +34,62 @@ Default.args = {
   options: [
     {
       value: 'BABYBIBXA19D9D000000XXXX',
+      icon: (
+        <Avatar
+          size='small'
+          src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BABYBIBXA19D9D000000XXXX_FLAT.png'
+          alt='Gray Baby Bib with Black Logo'
+        />
+      ),
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            alignIcon='center'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BABYBIBXA19D9D000000XXXX_FLAT.png'
-                alt='Gray Baby Bib with Black Logo'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Gray Baby Bib with Black Logo
-              </Text>
-              <Text size='small' tag='div' variant='info'>
-                200g
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Gray Baby Bib with Black Logo
+          </Text>
+          <Text size='small' tag='div' variant='info'>
+            200g
+          </Text>
+        </div>
       )
     },
     {
       value: 'BASEBHAT000000FFFFFFXXXX',
+      icon: (
+        <Avatar
+          size='small'
+          src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BASEBHAT000000FFFFFFXXXX_FLAT.png'
+          alt='Black Baseball Hat with White Logo'
+        />
+      ),
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            alignIcon='center'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BASEBHAT000000FFFFFFXXXX_FLAT.png'
-                alt='Black Baseball Hat with White Logo'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Black Baseball Hat with White Logo
-              </Text>
-              <Text size='small' tag='div' variant='info'>
-                50g
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Black Baseball Hat with White Logo
+          </Text>
+          <Text size='small' tag='div' variant='info'>
+            50g
+          </Text>
+        </div>
       )
     },
     {
       value: 'HOODIEUL000000FFFFFFLXXX',
+      icon: (
+        <Avatar
+          size='small'
+          src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/HOODIEUL000000FFFFFFLXXX_FLAT.png'
+          alt='Black Unisex Lightweight Hoodie'
+        />
+      ),
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            alignIcon='center'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/HOODIEUL000000FFFFFFLXXX_FLAT.png'
-                alt='Black Unisex Lightweight Hoodie'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Black Unisex Lightweight Hoodie
-              </Text>
-              <Text size='small' tag='div' variant='info'>
-                150g
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Black Unisex Lightweight Hoodie
+          </Text>
+          <Text size='small' tag='div' variant='info'>
+            150g
+          </Text>
+        </div>
       )
     }
   ]
@@ -150,32 +119,22 @@ WithQuantity.args = {
   options: [
     {
       value: 'BABYBIBXA19D9D000000XXXX',
+      icon: (
+        <Avatar
+          size='small'
+          src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BABYBIBXA19D9D000000XXXX_FLAT.png'
+          alt='Gray Baby Bib with Black Logo'
+        />
+      ),
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            alignIcon='center'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BABYBIBXA19D9D000000XXXX_FLAT.png'
-                alt='Gray Baby Bib with Black Logo'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Gray Baby Bib with Black Logo
-              </Text>
-              <Text size='small' tag='div' variant='info'>
-                200g
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Gray Baby Bib with Black Logo
+          </Text>
+          <Text size='small' tag='div' variant='info'>
+            200g
+          </Text>
+        </div>
       ),
       quantity: {
         min: 2,
@@ -184,32 +143,22 @@ WithQuantity.args = {
     },
     {
       value: 'BASEBHAT000000FFFFFFXXXX',
+      icon: (
+        <Avatar
+          size='small'
+          src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BASEBHAT000000FFFFFFXXXX_FLAT.png'
+          alt='Black Baseball Hat with White Logo'
+        />
+      ),
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            alignIcon='center'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/BASEBHAT000000FFFFFFXXXX_FLAT.png'
-                alt='Black Baseball Hat with White Logo'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Black Baseball Hat with White Logo
-              </Text>
-              <Text size='small' tag='div' variant='info'>
-                50g
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Black Baseball Hat with White Logo
+          </Text>
+          <Text size='small' tag='div' variant='info'>
+            50g
+          </Text>
+        </div>
       ),
       quantity: {
         min: 1,
@@ -218,32 +167,22 @@ WithQuantity.args = {
     },
     {
       value: 'HOODIEUL000000FFFFFFLXXX',
+      icon: (
+        <Avatar
+          size='small'
+          src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/HOODIEUL000000FFFFFFLXXX_FLAT.png'
+          alt='Black Unisex Lightweight Hoodie'
+        />
+      ),
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            alignIcon='center'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                src='https://res.cloudinary.com/commercelayer/image/upload/f_auto,b_white/demo-store/skus/HOODIEUL000000FFFFFFLXXX_FLAT.png'
-                alt='Black Unisex Lightweight Hoodie'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Black Unisex Lightweight Hoodie
-              </Text>
-              <Text size='small' tag='div' variant='info'>
-                150g
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Black Unisex Lightweight Hoodie
+          </Text>
+          <Text size='small' tag='div' variant='info'>
+            150g
+          </Text>
+        </div>
       ),
       quantity: {
         min: 1,
@@ -253,20 +192,11 @@ WithQuantity.args = {
     {
       value: 'TOTEBAGLE7DDC7000000XXXX',
       content: (
-        <>
-          <ListItem
-            tag='div'
-            borderStyle='none'
-            alignItems='center'
-            padding='none'
-          >
-            <div>
-              <Text size='regular' tag='div' weight='bold'>
-                Large Eco Tote Bag with Black Logo
-              </Text>
-            </div>
-          </ListItem>
-        </>
+        <div>
+          <Text size='regular' tag='div' weight='bold'>
+            Large Eco Tote Bag with Black Logo
+          </Text>
+        </div>
       ),
       quantity: {
         min: 1,
@@ -275,35 +205,28 @@ WithQuantity.args = {
     },
     {
       value: 'DHL',
+      icon: (
+        <Avatar
+          size='small'
+          shape='circle'
+          border='none'
+          src='carriers:dhl'
+          alt='DHL'
+        />
+      ),
       content: (
         <>
-          <ListItem
-            tag='div'
-            alignIcon='center'
-            borderStyle='none'
-            padding='none'
-            icon={
-              <Avatar
-                size='small'
-                shape='circle'
-                border='none'
-                src='carriers:dhl'
-                alt='DHL'
-              />
-            }
-          >
-            <div>
-              <Text size='regular' weight='bold'>
-                Express Easy
-              </Text>
-              <Text size='small' tag='div' variant='info' weight='medium'>
-                DHL Express
-              </Text>
-            </div>
-            <Text size='regular' weight='bold' wrap='nowrap'>
-              €37,61
+          <div>
+            <Text size='regular' weight='bold'>
+              Express Easy
             </Text>
-          </ListItem>
+            <Text size='small' tag='div' variant='info' weight='medium'>
+              DHL Express
+            </Text>
+          </div>
+          <Text size='regular' weight='bold' wrap='nowrap'>
+            €37,61
+          </Text>
         </>
       ),
       quantity: {
